# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InterlinedList API",
    "version": "0.1.0",
    "description": "HTTP API for InterlinedList. Generated deterministically from the route tree under app/api/ and enriched from docs/api-reference.md by scripts/generate-openapi.js (`npm run docs:openapi`). Served at /api/openapi.json and rendered at /api-docs.\n\nMost endpoints accept either a session cookie (`interlinedlist-session`) or a Bearer sync token from POST /api/auth/sync-token. A subset (auth account management, exports, identities, organizations, LinkedIn, message dig, architecture aggregates) requires the session cookie."
  },
  "servers": [
    {
      "url": "https://interlinedlist.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/messages/search": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Search top-level messages by content (respecting feed visibility)",
        "operationId": "getMessagesSearch",
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "content": {
                            "type": "string"
                          },
                          "publiclyVisible": {
                            "type": "boolean"
                          },
                          "imageUrls": {
                            "type": "array",
                            "items": {}
                          },
                          "videoUrls": {
                            "type": "array",
                            "items": {}
                          },
                          "tags": {
                            "type": "array",
                            "items": {}
                          },
                          "crossPostUrls": {
                            "type": "array",
                            "items": {}
                          },
                          "scheduledAt": {},
                          "pushedMessageId": {},
                          "pushCount": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "username": {
                                "type": "string"
                              },
                              "displayName": {
                                "type": "string"
                              },
                              "avatar": {}
                            }
                          },
                          "dugByMe": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  },
                  "example": {
                    "messages": [
                      {
                        "id": "msg123",
                        "content": "Hello world from the API",
                        "publiclyVisible": true,
                        "imageUrls": [],
                        "videoUrls": [],
                        "tags": [],
                        "crossPostUrls": [],
                        "scheduledAt": null,
                        "pushedMessageId": null,
                        "pushCount": 0,
                        "createdAt": "2026-06-04T10:00:00.000Z",
                        "updatedAt": "2026-06-04T10:00:00.000Z",
                        "user": {
                          "id": "u1",
                          "username": "alice",
                          "displayName": "Alice",
                          "avatar": null
                        },
                        "dugByMe": false
                      }
                    ],
                    "pagination": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0,
                      "hasMore": false
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Sync token from POST /api/auth/sync-token, sent as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "interlinedlist-session",
        "description": "Session cookie set by POST /api/auth/login."
      }
    },
    "parameters": {
      "limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 50
        },
        "description": "Page size."
      },
      "offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        },
        "description": "Items to skip."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Not authenticated.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Not authenticated"
        }
      },
...
    }
  }
}